+++ src/Symfony/Component/DependencyInjection/Definition.php
@@ -30,7 +30,11 @@ class Definition
     private $properties = array();
     private $calls = array();
     private $getters = array();
+<<<<<<< HEAD
     private $instanceof = array();
+=======
+    private $tails = array();
+>>>>>>> FETCH_HEAD
     private $configurator;
     private $tags = array();
     private $public = true;
@@ -39,6 +43,7 @@ class Definition
     private $lazy = false;
     private $decoratedService;
     private $autowiredCalls = array();
+    private $autowiredTails = array();
     private $autowiringTypes = array();
 
     protected $arguments;
@@ -365,6 +370,7 @@ class Definition
     }
 
     /**
+<<<<<<< HEAD
      * Sets the definition templates to conditionally apply on the current definition, keyed by parent interface/class.
      *
      * @param $instanceof ChildDefinition[]
@@ -374,11 +380,24 @@ class Definition
     public function setInstanceofConditionals(array $instanceof)
     {
         $this->instanceof = $instanceof;
+=======
+     * @return $this
+     *
+     * @experimental in version 3.3
+     */
+    public function setOverridenTail($name, array $defaultArgs)
+    {
+        if (!$name) {
+            throw new InvalidArgumentException(sprintf('Tail method name cannot be empty.'));
+        }
+        $this->tails[strtolower($name)] = $defaultArgs;
+>>>>>>> FETCH_HEAD
 
         return $this;
     }
 
     /**
+<<<<<<< HEAD
      * Gets the definition templates to conditionally apply on the current definition, keyed by parent interface/class.
      *
      * @return ChildDefinition[]
@@ -388,6 +407,25 @@ class Definition
     public function getInstanceofConditionals()
     {
         return $this->instanceof;
+=======
+     * @return $this
+     *
+     * @experimental in version 3.3
+     */
+    public function setOverridenTails(array $tails)
+    {
+        $this->tails = array_change_key_case($tails, CASE_LOWER);
+
+        return $this;
+    }
+
+    /**
+     * @experimental in version 3.3
+     */
+    public function getOverridenTails()
+    {
+        return $this->tails;
+>>>>>>> FETCH_HEAD
     }
 
     /**
@@ -737,7 +775,7 @@ class Definition
      */
     public function isAutowired()
     {
-        return !empty($this->autowiredCalls);
+        return $this->autowiredCalls || $this->autowiredTails;
     }
 
     /**
@@ -779,6 +817,26 @@ class Definition
     public function setAutowiredCalls(array $autowiredCalls)
     {
         $this->autowiredCalls = $autowiredCalls;
+
+        return $this;
+    }
+
+    /**
+     * @experimental in 3.3
+     */
+    public function getAutowiredTails()
+    {
+        return $this->autowiredTails;
+    }
+
+    /**
+     * @experimental in 3.3
+     *
+     * @return $this
+     */
+    public function setAutowiredTails(array $autowiredTails)
+    {
+        $this->autowiredTails = $autowiredTails;
 
         return $this;
     }