+++ src/Symfony/Component/DependencyInjection/Dumper/GraphvizDumper.php
@@ -87,6 +87,13 @@ class GraphvizDumper extends Dumper
                     $this->findEdges($id, $value, false, $name.'()')
                 );
             }
+
+            foreach ($definition->getOverridenTails() as $name => $value) {
+                $this->edges[$id] = array_merge(
+                    $this->edges[$id],
+                    $this->findEdges($id, $value, false, $name.'()')
+                );
+            }
         }
 
         return $this->container->resolveEnvPlaceholders($this->startDot().$this->addNodes().$this->addEdges().$this->endDot(), '__ENV_%s__');