+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -4,7 +4,11 @@ CHANGELOG
 3.3.0
 -----
 
+<<<<<<< HEAD
  * [EXPERIMENTAL] added "instanceof" section for local interface-defined configs
+=======
+ * [EXPERIMENTAL] added support for tail-injection for methods + autowiring
+>>>>>>> FETCH_HEAD
  * [EXPERIMENTAL] added "service-locator" argument for lazy loading a set of identified values and services
  * [EXPERIMENTAL] added prototype services for PSR4-based discovery and registration
  * added `ContainerBuilder::getReflectionClass()` for retrieving and tracking reflection class info